Fix the database bug at "Delete OSCAR Clients"
git-svn-id: http://svn.oscar.openclustergroup.org/oscar/trunk@4095 18f9d06c-6ae2-0310-8438-938a554feff9

Former-commit-id: e5b9c4a30d5599d1a087d276d734ce125819a57f [formerly 35774393d9ce71402fdab1024e152dd1170979ee] [formerly 6bf44096040c27c8b2c7cd48c87e58144339dbf1 [formerly 5b1e552c0e46f83215de52c5c0767aec6a251477]] [formerly 6364d5c9cd0e8634936c001e7b3f8d16cc8e79e1 [formerly 2ebbf4318490ec96f5e7fe6b15f0653b726bcce2] [formerly 6e2c09fd49202cefdb090d7139fca5cb0d006874 [formerly f58c231f215df3d14542ba3d64891598c51ddbfd]]]
Former-commit-id: 0e858e87224b99a5592f5d05d236d96f55aa9e91 [formerly b57d9fee1fdab8fbfac3db97f80a429d14e5ca06] [formerly 7a17a571aac59559e80e074e036c03caab212b39 [formerly b1deddf4016855814159d6295f0dc710ba9514ab]]
Former-commit-id: eabcd8e100bf0f27521cd918ce94e1ca324f23d6 [formerly f86c2073e344520acca6328c168edef293034c44]
Former-commit-id: 02a32bc9309877a2b9de9bf77c6f6f9ca7009456

+++ lib/OSCAR/Database.pm
@@ -63,6 +63,7 @@ if(-e '/etc/odapw'){
               update_table
               select_table
               delete_table
+              delete_node
               get_client_nodes
               get_networks
               get_packages_related_with_package
@@ -1514,11 +1515,39 @@ sub delete_node{
     my ($node_name,
         $options_ref,
         $error_strings_ref) = @_;
+    my $node_ref = get_node_info_with_name($node_name,$options_ref,$error_strings_ref);    
+    my $node_id = $$node_ref{id};
+    return 1 if !$node_id;
+
+    delete_group_node($node_id,$options_ref,$error_strings_ref);
+    delete_node_packages($node_id,$options_ref,$error_strings_ref);
     my $sql = "DELETE FROM Nodes ";
     $sql .= ($node_name?"WHERE name='$node_name'":"");
     die "$0:Failed to update values via << $sql >>"
         if! do_update($sql,"Nodes", $options_ref, $error_strings_ref);
 }    
+
+sub delete_group_node{
+    my ($node_id,
+        $options_ref,
+        $error_strings_ref) = @_;
+    my $sql = "DELETE FROM Group_Nodes WHERE node_id=$node_id";
+    die "$0:Failed to update values via << $sql >>"
+        if! do_update($sql,"Group_Nodes", $options_ref, $error_strings_ref);
+}
+
+sub delete_node_packages{
+    my ($node_id,
+        $options_ref,
+        $error_strings_ref) = @_;
+    my $sql = "DELETE FROM Node_Packages WHERE node_id=$node_id";
+    die "$0:Failed to update values via << $sql >>"
+        if! do_update($sql,"Node_Packages", $options_ref, $error_strings_ref);
+    $sql = "DELETE FROM Node_Package_Status WHERE node_id=$node_id";
+    die "$0:Failed to update values via << $sql >>"
+        if! do_update($sql,"Node_Package_Status", $options_ref, $error_strings_ref);
+}
+
 
 sub get_client_nodes_info{
     my ($server,
@@ -1743,13 +1772,15 @@ sub get_packages_switcher{
 
 sub get_packages_servicelists{
     my ($results_ref,
-        $name,
+        $group_name,
         $options_ref,
         $error_strings_ref) = @_;
-    my $sql = "SELECT P.package, S.service " .
-              "FROM Packages P, Packages_servicelists S " .
-              "WHERE P.id=S.package_id ";
-    $sql .= ($name?" AND P.package='$name'":"");          
+    my $sql = "SELECT distinct P.package, S.service " .
+              "FROM Packages P, Packages_servicelists S, Node_Package_Status N, " .
+              "Group_Nodes G " .
+              "WHERE P.id=S.package_id AND N.package_id=S.package_id ".
+              "AND G.node_id=N.node_id ";
+    $sql .= ($group_name?" AND G.group_name='$group_name' AND S.group_name='$group_name'":" AND S.group_name!='oscar_server'");          
     return do_select($sql,$results_ref,$options_ref,$error_strings_ref);
 }
